Split view.hpp into range_view.hpp & all.hpp
