+++ agency/experimental/ranges/range_view.hpp
@@ -1,11 +1,6 @@
 #pragma once
 
-// XXX we should rename this file something like range_view.hpp
-//     and move all() into its own header
-
 #include <agency/detail/config.hpp>
-#include <agency/experimental/array.hpp>
-#include <agency/experimental/span.hpp>
 #include <agency/experimental/ranges/range_traits.hpp>
 #include <type_traits>
 #include <iterator>
@@ -14,49 +9,6 @@ namespace agency
 {
 namespace experimental
 {
-
-
-// XXX this is only valid for contiguous containers
-template<class Container>
-__AGENCY_ANNOTATION
-span<typename Container::value_type> all(Container& c)
-{
-  return span<typename Container::value_type>(c);
-}
-
-// XXX this is only valid for contiguous containers
-template<class Container>
-__AGENCY_ANNOTATION
-span<const typename Container::value_type> all(const Container& c)
-{
-  return span<const typename Container::value_type>(c);
-}
-
-
-template<class T, std::size_t N>
-__AGENCY_ANNOTATION
-span<T,N> all(array<T,N>& a)
-{
-  return span<T,N>(a);
-}
-
-
-template<class T, std::size_t N>
-__AGENCY_ANNOTATION
-span<const T,N> all(const array<T,N>& a)
-{
-  return span<const T,N>(a);
-}
-
-
-// spans are already views, so don't wrap them
-// XXX maybe should put this in span.hpp
-template<class ElementType, std::ptrdiff_t Extent>
-__AGENCY_ANNOTATION
-span<ElementType,Extent> all(span<ElementType,Extent> s)
-{
-  return s;
-}
 
 
 template<class Iterator, class Sentinel = Iterator>
